rename repo from atlas to dolphin
