+++ setup.cfg
@@ -1,10 +1,10 @@
 [metadata]
-name = atlas
+name = dolphin
 version = 0.1.0
 description = Workflows for generating surface displacement maps using InSAR
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/opera-adt/atlas
+url = https://github.com/opera-adt/dolphin
 classifiers =
     Development Status :: 2 - Pre-Alpha
     Intended Audience :: Science/Research
@@ -31,12 +31,12 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    atlas = atlas.s1_disp:main
+    dolphin = dolphin.s1_disp:main
 
 
 [options.extras_require]
 # Add here additional requirements for extra features, to install with:
-# `pip install atlas[dev]`
+# `pip install dolphin[dev]`
 dev =
     pytest
     pre-commit
