+++ src/dolphin/defaults/s1_disp_stack.yaml
