+++ src/dolphin/phase_linking.py
@@ -3,7 +3,7 @@ import os
 from os import fspath
 from pathlib import Path
 
-from atlas.utils import Pathlike, copy_projection
+from dolphin.utils import Pathlike, copy_projection
 
 # nmap.py -i stack/slcs_base.vrt -o nmap/nmap -c nmap/count -x 11 -y 5
