+++ src/dolphin/s1_disp_single.py
@@ -2,7 +2,7 @@
 # from pathlib import Path
 
 # from atlas import combine_ps_ds, phase_linking, ps  # , unwrap
-from atlas.log import get_log, log_runtime
+from dolphin.log import get_log, log_runtime
 
 logger = get_log()
