+++ src/dolphin/schemas/s1_disp_stack.yaml
