+++ tests/test_config.py
@@ -4,11 +4,11 @@ from pathlib import Path
 import pytest
 import yamale
 
-import atlas.config
-from atlas.config import get_workflow_yaml_path
+import dolphin.config
+from dolphin.config import get_workflow_yaml_path
 
 
-# Using fixtures for these to avoid nagivating the paths from tests/ to src/
+# Using fixtures for these to avoid navigating the paths from tests/ to src/
 @pytest.fixture
 def schema_files():
     return [
@@ -42,7 +42,7 @@ def test_yaml_stack():
     defaults_file = get_workflow_yaml_path("s1_disp_stack.yaml", yaml_type="defaults")
     # Check that the updating of defaults works
     minimal_path = Path(__file__).parent / "data/s1_disp_stack_minimal.yaml"
-    min_data = atlas.config.load_workflow_yaml(
+    min_data = dolphin.config.load_workflow_yaml(
         minimal_path, workflow_name="s1_disp_stack"
     )
     min_proc_dict = min_data["runconfig"]["groups"]["processing"]
@@ -59,24 +59,24 @@ def test_yaml_save(tmp_path):
     from ruamel.yaml import YAML
 
     minimal_path = Path(__file__).parent / "data/s1_disp_stack_minimal.yaml"
-    min_data = atlas.config.load_workflow_yaml(
+    min_data = dolphin.config.load_workflow_yaml(
         minimal_path, workflow_name="s1_disp_stack"
     )
 
     temp_file = tmp_path / "out.yaml"
-    atlas.config.save_yaml(temp_file, min_data)
+    dolphin.config.save_yaml(temp_file, min_data)
     y = YAML()
     assert y.load(temp_file) == min_data
 
 
-def test_atlas_cfg_section():
+def test_dolphin_cfg_section():
     minimal_path = Path(__file__).parent / "data/s1_disp_stack_minimal.yaml"
-    min_data = atlas.config.load_workflow_yaml(
+    min_data = dolphin.config.load_workflow_yaml(
         minimal_path, workflow_name="s1_disp_stack"
     )
-    cfg_augmented = atlas.config.add_atlas_section(min_data)
+    cfg_augmented = dolphin.config.add_dolphin_section(min_data)
 
     now = str(datetime.datetime.now())
     # '2022-08-03 08:47:28.834660' , ignore microseconds
     proc_dict = cfg_augmented["runconfig"]["groups"]["processing"]
-    assert proc_dict["atlas"]["runtime"][:-7] == now[:-7]
+    assert proc_dict["dolphin"]["runtime"][:-7] == now[:-7]
